requirements.txt
requests==2.32.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
aiohttp==3.12.15